requirements.txt
llama-index==0.10.18
pymupdf
pinecone-client
python-dotenv
